Move map to restaurant activity
