+++ Restervator/app/build.gradle
@@ -33,6 +33,7 @@ android {
 
 dependencies {
     implementation 'com.jakewharton:butterknife:10.2.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     annotationProcessor 'com.jakewharton:butterknife-compiler:10.2.0'
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
