+++ Restervator/app/src/main/java/com/restervator/RestaurantActivity.java
@@ -1,7 +1,9 @@
 package com.restervator;
 
+import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.preference.PreferenceManager;
 import android.widget.ImageView;
 
 import androidx.appcompat.app.AppCompatActivity;
@@ -10,13 +12,57 @@ import androidx.appcompat.widget.Toolbar;
 import com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton;
 import com.squareup.picasso.Picasso;
 
+import org.osmdroid.api.IMapController;
+import org.osmdroid.config.Configuration;
+import org.osmdroid.tileprovider.tilesource.TileSourceFactory;
+import org.osmdroid.util.GeoPoint;
+import org.osmdroid.views.MapView;
+import org.osmdroid.views.overlay.mylocation.GpsMyLocationProvider;
+import org.osmdroid.views.overlay.mylocation.MyLocationNewOverlay;
+
 
 public class RestaurantActivity extends AppCompatActivity {
+
+
+    MapView map = null;
+    private MyLocationNewOverlay mLocationOverlay;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_restaurant);
+
+        /*******************Loading the map ***************/
+
+        //load/initialize the osmdroid configuration for the map
+        Context ctx = getApplicationContext();
+        Configuration.getInstance().setUserAgentValue(BuildConfig.APPLICATION_ID);
+        Configuration.getInstance().load(ctx, PreferenceManager.getDefaultSharedPreferences(ctx));
+
+        //setting this before the layout is inflated is a good idea
+        //it 'should' ensure that the map has a writable location for the map cache, even without permissions
+        //note, the load method also sets the HTTP User Agent to your application's package name, abusing osm's tile servers will get you banned based on this string
+        //inflate and create the map
+        setContentView(R.layout.activity_restaurant);
+        map = findViewById(R.id.map);
+        map.setTileSource(TileSourceFactory.MAPNIK);
+        map.setMultiTouchControls(true);
+        IMapController mapController = map.getController();
+        mapController.setZoom(9.5);
+
+
+        //adds the user location on the map. This is not yet working, right now default location is set to milan below
+//        this.mLocationOverlay = new MyLocationNewOverlay(new GpsMyLocationProvider(ctx), map);
+//        this.mLocationOverlay.enableMyLocation();
+//        map.getOverlays().add(this.mLocationOverlay);
+
+        //sets the starting point on the map to the long and lat of Milan
+        GeoPoint startPoint = new GeoPoint(45.4642, 9.1900);
+        mapController.setCenter(startPoint);
+
+
+        /*******************Fetching restaurant image, name, and Fab controls ***************/
+
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
 
@@ -34,5 +80,21 @@ public class RestaurantActivity extends AppCompatActivity {
                 .load("https://media-cdn.tripadvisor.com/media/photo-s/0e/cc/0a/dc/restaurant-chocolat.jpg")
                 .into(restaurantImageView);
     }
+
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        //this will refresh the osmdroid configuration on resuming.
+        map.onResume();
+    }
+
+    @Override
+    public void onPause() {
+        super.onPause();
+        //this will refresh the osmdroid configuration on resuming.
+        map.onPause();
+    }
+
 
 }
