+++ Restervator/app/src/main/res/values/strings.xml
@@ -28,4 +28,8 @@
         please enable the permissions to use this app.
     </string>
     <string name="List_image">firstimage</string>
+    <string name="more_information">More information</string>
+
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
